Work around gpg --trust-model=always bug
When you specify --trust-model=always to gpg2, it doesn't create the
trustdb if it doesn't already exist. Alas, some gpg2 commands expect
the trustdb to be there even when --trust-model=always is specified.
To work around this problem, Don't specify --trust-model=always when
generating the initial server and client keys, because then the
trustdb will be created at that point and everything will work after
that.