Use RSS feed for Arch Linux security patches, not mailing list archive
The mailing list archive for the Arch Linux security list is only
updated daily, so it's not a good choice for finding out about
security updates promptly.

Furthermore, the whole point of an RSS feed is to be queried
relatively frequently, so rather than only checking for patches once
per day, we should do it hourly.