Make GnuPG version checker more flexible
The code for checking what version of GnuPG is installed was assuming
that the version number would be on the first line of the output of
`gpg --version`. Unfortunately, it isn't always so. Most notably, when
we close stdin while daemonizing and GnuPG decides to "reopen" it as a
courtesy to us (gee, thanks!), it prints a warning about that as the
first line of the output.

Therefore, we know allow the version-number regular expression to
match anywhere in the output, not just on the first line.