Add remote shells support
Add the ability to initiate a remote shell request on the server which
is then responded to by the client the next time it checks in with the
server, resulting in an interactive shell connection between the
server and the client, proxied through the PenguinDome web server.

Making this work actually required enhancements and improvements in
many different areas, all of which are included in this bug commit
since it's not worth trying to tease them out into separate commits.
These include:

* The client now requires the additional Python packages `pexpect`,
  `pycrypto`, and `python-daemon`.

* Script communication with the server now uses a session so that
  keep-alive works and (when appropriate) authentication is preserved
  between requests.

* The server has been fixed so that keep-alive works.

* The web server now listens on a separate port only on 127.0.0.1,
  intended only for the use of scripts and utilities running locally
  on the same machine as the web server. The port number for this is
  configured via the `local_port` setting in `server/settings.yml`.

* GnuPG signing can now be turned off in server requests when it's
  appropriate to do so.

* Users (with passwords) and groups can now be configured at a
  server-wide level in `server/settings.yml`.