For not-reporting check, support the concept of clients linked by user
For the not-reporting check in issues.py, support the idea that if a
user has more than one client they use, it's OK for some of their
clients not to have reported recently if others have. As a backstop to
this, however, all clients linked in this way need to check in at
least once per month.

To make this possible, the concept of "client parameters" has been
added, with a new script, server/client_parameters.py, for
setting, unsetting, and listing client parameters.