Nothing should barf because of race conditions in psutil.process_iter()
I've fixed bits and pieces of this problem in previous commits, but
this is an attempt to comprehensively address all of them.

`psutil.process_iter()` sometimes returns processes that go away
between when it is called and when we try to access their attributes,
causing a `FileNotFoundError` exception.

To avoid this problem, create a new helper method,
`penguindome.plugin_tools.process_dict_iter()`, which takes a list of
attributes that are wanted from each process and returns a dict of
processes that have all those attributes. When this iterator
encounters `FileNotFoundError` internally while iterating, it skips to
the next process.

Alas, this helper method can't be used everywhere. The `sshd` plugin
needs to fetch data about processes that `Process.as_dict()` doesn't
support, so that plugin needs to use `psutil.process_iter()` directly.
This commit therefore also adds logic to protect that plugin's code
from the `FileNotFoundError` race condition.