Make submit quieter about known failure modes
Get rid of some log noise by making submit quieter about known failure
modes, i.e., HTTPError, ReadTimeout. It still logs the full details of
the failure in the debug logs, but the error logs only get a one-line
log message about the failure, so there's much less noise in the logs
that we should be paying attention to in real-time.