Make signature verification failures quieter and more informative
Log as info rather than error when the signature verification code
tries to verify the signature in an empty request.

When a non-empty request has no signature or has an invalid signature,
log it as an error rather than logging a full stack trace.

Try to include the remote address and host name in log messages about
signature verification failures.