Use utf8 everywhere instead of ascii for decoding command output
Some subprocess commands we execute may produce non-ASCII output, most
likely UTF-8. I noticed this specifically with the output of `gpg
--verify` on a system using a non-English locale such that there were
UTF-8 characters in the month name (`gpg --verify` prints the date
that the signature being verified was created).

Since UTF-8 is a superset of ASCII, it's safe to just try to decode
with utf8 everywhere rather than ascii.