Change how server runs to fix socket CLOSE_WAIT problem
Werkzeug has a bug which causes sockets to get stuck in TCP CLOSE_WAIT
state. Eventually, there are so many stuck sockets that the kernel
stops allowing new connections to the port, and then attempts to
connect to the port by clients hang. The "real" way to fix this is to
run the server under a different WSGI framework, but all of them are
more of a pain to set up than werkzeug, so I'm hoping that we can
solve this problem just by handling each request in a separate process
so the socket for each request will get closed properly when its
process exits.

More info:
https://stackoverflow.com/questions/31403261/flask-werkzeug-sockets-stuck-in-close-wait

As part of this change, I'm upgrading the server PyPI packages to
current versions. I'm also getting rid of the setting controlling
whether to run the server in threaded mode, because with this change
we're always running it in multi-process mode. If you need to change
that during development or debugging, you can change the 'processes'
setting to 1 in `startServer` in server/server.py.