BUG: PID wraparound could cause wrong xinit uesr to be identified
We were using the smallest PID with a given DISPLAY value to determine
the owner of an X display. If the user's PID's wrap around and the
lowest-numbered running process with their DISPLAY setting is owned by
a user other than them, that'll break. So now we use the process with
the lowest creation time instead.