Fix client shell, broken in ae1a3d2
The method that I used to specify the GPG user ID explicitly in commit
ae1a3d2 was incorrect because it assumed that all API invocations came
from the client, when in fact the server makes an API invocation when
a user on the server initiates a client shell. This commit fixes the
method so that so the correct GPG user ID Is used when the API is
invoked on the server.

Fixes quantopian/PenguinDome#19.