Upgrade PyYAML to address deprecation warnings, security issues
The `yaml.load` function is now deprecated without specifying a loader
explicitly, and using the full loader -- which has a larger attack
surface -- is discouraged when it isn't needed. Furthermore, the older
version of `PyYAML` we were using generates unrelated deprecation
warnings in Python 3.8.

For these reasons, upgrade to the current version of PyYAML and switch
from using `yaml.load` to `yaml.safe_load`.