Update versions of all packages in requirements files
Full disclosure: the test coverage in this project is minimal and I
haven't actually tried running the application with these new package
versions. I am not currently using it in production so I don't have
any need to do that, but I wanted to get updated versions committed
for security-update reasons. If things don't work for you, please file
an issue and I will investigate and fix (or if you have the bandwidth
investigate yourself and file a PR to fix it!). I'm not going to spend
more time working on this unless there's somebody actually using this
package who would benefit from the work.