Use my patched version of the `stopit` package
The current version of the `stopit` package on PyPI is incompatible
with Python 3.12 because it uses `pkg_resources`, which is no longer
available for import by default. I'm therefore switching to a patched
version that uses `importlib` instead. If/when the maintainers of
stopit accept my pull request and release a new version with this
change in it, I will switch back to using the PyPI version.