Switch to my patched version of openedx's MongoDBProxy
Openedx's MongoDBProxy uses the `pyandoc` package in its setup script.
The way they're using it is incompatible with the current version of
the package, and in any case the package is completely deprecated, and
in any case it's no longer necessary for us to be doing pandoc
processing at all on the long description in setup.py because
setuptools handles that all automatically now, so my patched version
takes advantage of that. If/when they accept my patch I'll switch back
to using the openedx version of the package.